requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
Django==6.0.4
django-axes==8.3.1
djangorestframework==3.17.1
et_xmlfile==2.0.0
gunicorn==25.3.0
idna==3.12
lxml==6.1.0
openpyxl==3.1.5
packaging==26.1
pillow==12.2.0
psycopg2-binary==2.9.12
PyJWT==2.12.1
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.2
reportlab==4.4.10
requests==2.33.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
whitenoise==6.12.0
